+++ examples/oft-hyperliquid/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-hyperliquid-example
 
+## 1.0.10
+
+### Patch Changes
+
+- bb2f4c3: sdk updates for quote asset improvement
+
+## 1.0.9
+
+### Patch Changes
+
+- bfea443: alignedQuoteAsset support on core spot deployment + instructions on how to create a custom hyperliquid sdk action for sign + submit
+
 ## 1.0.8
 
 ### Patch Changes